ref(spans): Keep mobile constants in normalize::utils
Revert the constant move — the deprecated comment on utils.rs is 3
years old and the module is still actively used. Keep MOBILE_SDKS,
MAIN_THREAD_NAME, and MAX_DURATION_MOBILE_MS where they are.
