+++ relay-event-normalization/src/normalize/utils.rs
@@ -8,6 +8,23 @@ use std::f64::consts::SQRT_2;
 use relay_event_schema::protocol::{Event, ResponseContext, Span, TraceContext, User};
 use relay_protocol::Value;
 
+/// Used to decide when to extract mobile-specific tags.
+pub const MOBILE_SDKS: [&str; 4] = [
+    "sentry.cocoa",
+    "sentry.dart.flutter",
+    "sentry.java.android",
+    "sentry.javascript.react-native",
+];
+
+/// Allowed value for main thread name.
+pub const MAIN_THREAD_NAME: &str = "main";
+
+/// Maximum duration of a mobile measurement in milliseconds.
+///
+/// Mobile measurements (app start, TTID, TTFD) that exceed this threshold are considered
+/// outliers and removed.
+pub const MAX_DURATION_MOBILE_MS: f64 = 180_000.0;
+
 /// Extract the HTTP status code from the span data.
 pub fn http_status_code_from_span(span: &Span) -> Option<String> {
     // For SDKs which put the HTTP status code into the span data.
